+++ test/resources/checkers/perl-error.pl
@@ -1,4 +1,0 @@
-# An unused variable
-
-use warnings;
-$x = 10;