+++ test/resources/checkers/perl-syntax-error.pl
@@ -1,5 +1,0 @@
-# A syntax error
-
-use strict
-use warnings;
-$x = 10;