+++ test/resources/checkers/perl.pl
@@ -1,0 +1,11 @@
+package perl;
+
+use strict;
+use warnings;
+
+$x = <*.pl>;
+
+system('ls');
+
+die if undef;
+1;